+++ GAC-EJBFacade/ejbModule/remote/FProjectServicesRemote.java
@@ -8,7 +8,7 @@ import model.Project;
 public interface FProjectServicesRemote {
 	 public void addItem(Project i);
 
-	    public Project findItem(long id) ;
+	    public Project findItem(Integer id) ;
 
 	    public void deleteItem(Project i) ;
